Add disqus threads and TOCs to doc pages. A few syntax style tweaks.
The syntax tweaks are to hopefully make the github rendering of the
markdown pages a little easier to understand. The frontmatter now
doesn't get rendered as an H1 on github, and we'll use mustache
variables for the very simple variables on the page, which might make
more sense to non-ruby people.