Switch to Middleman from nanoc for static site compiling:
This is in semi preparation for redoing the data.gov theme, but this was
also a cleanup I've been wanting to do for a while for a few reasons,
but mainly:

- A lot of the nanoc customizations are simply baked into middleman, so
it cleans up various things (asset pipeline, syntax highlighting, etc).
- Middleman's live server preview is quite nice to work with when
developing new content, so hopefully this will make it easier for other
users to work against this project locally.