Introduce the concept of email verification to signup process.
It's simple, but we're targeting a backwards-compatible optional way.
During the signup process, if email verification is enabled, then the
api key won't be returned or displayed and will instead be e-mailed to
the user. We'll assume if the API key is used, then the e-mail address
is verified.

Conflicts:
	source/static/javascripts/_signup_embed/setup.js.erb