Add missing agencies to our documentation list.
Also redo how the /docs/ listing is pulled together yet again. It's back
to being dynamic, because I'm exceedingly lazy. So it's a little bit
magical, but it's consistent with how the home page was automatically
making it's listing. It also simplifies things so we just have to
maintain the individual agency pages, so hopefully this works.