Fix Node.js version used in GitHub actions.
We were never explicitly configuring the Node.js version, so with a
recent update to 16, that broke some of our builds. So explicitly
install v14 with the "setup-node" action.

Also move to the "ruby/setup-ruby" action for installing ruby, since the
previous action was deprecated.

Leverage the dependency caching built into both of these actions, which
should eliminate some of our custom caching setup.