Trigger change in built embed css file to clear caches.
Since the temporary workaround implemented in
85fba3aabb6dcf662166216fa02bd4dd3fc6a86a happens after the build, this
didn't trigger a fresh version of the asset being deployed. So tweak
something else so that a new cache-busted asset gets deployed with this
change. This should all be temporary until our new signup form updates
go live, so while not ideal, this shouldn't be long-lived.